requirements.txt
vllm>=0.6.0
openai>=1.0.0
httpx>=0.27.0
pyyaml>=6.0.1
numpy>=1.26.0
tqdm>=4.66.0
